requirements.txt
websockets
uvloop
jamspell
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
cytoolz==0.9.0.1
Django==2.0.3
django-appconf==1.0.2
django-filter==1.1.0
django-user-accounts==2.0.3
eth-abi==0.5.0
eth-keys==0.1.0b4
eth-tester==0.1.0b11
eth-utils==0.8.1
gunicorn==19.7.1
idna==2.6
jsonfield==2.0.2
kombu==4.1.0
lru-dict==1.1.6
nltk==3.2.5
numpy==1.14.2
Pillow==5.0.0
py-solc==2.1.0
pysha3==1.0.2
pytz==2018.3
requests==2.18.4
rlp==0.6.0
scikit-learn==0.19.1
scipy==1.0.0
semantic-version==2.6.0
six==1.11.0
sklearn==0.0
toolz==0.9.0
urllib3==1.22
vine==1.1.4
web3==3.16.5
mysqlclient==1.3.12
